pyproject.toml
[tool.poetry]
name = "bigbedbot-py"
version = "0.1.0"
description = ""
authors = ["mikezom <mikezom3721@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
graia-ariadne = {extras = ["full"], version = "^0.7.6"}
prompt-toolkit = "^3.0.29"
graia-saya = "^0.0.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false